requirements.txt
langchain_groq
langchain
python-dotenv
langchain_community
langchain_huggingface
langchain_chroma
langchain_core
sentence-transformers
pypdf
streamlit